Genericise renderArgs function, removing knowledge of slices from the menu items & use generic one in ftrace and counter tracks details panels. Created renderSliceArguments() which has the additional menu items
