+++ ui/src/components/sql_utils/args.ts
@@ -102,8 +102,7 @@ function parseValue(
     case 'string':
       return value.stringValue;
     case 'bool':
-      // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
-      return !!value.intValue;
+      return value.intValue === null ? null : value.intValue !== 0n;
     case 'real':
       return value.realValue;
     case 'null':
